refactor(Car): create racing.car pkg
