--- src/main/java/racing/MoveChanceGenerator.java
@@ -1,6 +1,0 @@
-package racing;
-
-@FunctionalInterface
-interface MoveChanceGenerator {
-	int generateChance();
-}
