+++ src/main/java/racing/RacingGame.java
@@ -1,5 +1,7 @@
 package racing;
 
+import racing.car.Car;
+
 import java.util.List;
 import java.util.Random;
 import java.util.stream.Collectors;
