+++ src/main/java/racing/car/MoveChanceGenerator.java
@@ -1,0 +1,6 @@
+package racing.car;
+
+@FunctionalInterface
+public interface MoveChanceGenerator {
+	int generateChance();
+}
